+++ src/main/java/de/dddns/kirbylink/warp4j/model/Target.java
@@ -1,0 +1,9 @@
+package de.dddns.kirbylink.warp4j.model;
+
+import lombok.Value;
+
+@Value
+public class Target {
+  Platform platform;
+  Architecture architecture;
+}