+++ src/test/java/de/dddns/kirbylink/warp4j/config/Warp4JCommandTest.java
@@ -8,15 +8,22 @@ import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.Set;
 import java.util.stream.Stream;
 import javax.naming.NoPermissionException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.MockedStatic;
+import de.dddns.kirbylink.warp4j.config.Warp4JCommand.Warp4JCommandConfiguration;
+import de.dddns.kirbylink.warp4j.model.Architecture;
+import de.dddns.kirbylink.warp4j.model.Platform;
+import de.dddns.kirbylink.warp4j.model.Target;
 import de.dddns.kirbylink.warp4j.service.Warp4JService;
 import de.dddns.kirbylink.warp4j.utilities.ComponentFactory;
 
@@ -73,5 +80,66 @@ class Warp4JCommandTest {
         Arguments.of(IOException.class, 74),
         Arguments.of(NoPermissionException.class, 77)
     );
+  }
+
+  @Nested
+  @DisplayName("Test for Warp4JCommandConfiguration")
+  class TestWarp4JCommandConfiguration {
+    @Test
+    void testGetAllSelectedTargets_WhenNothingIsSelected_ThenAllTargetsReturned() {
+      // Given
+      var warp4JCommandConfiguration = Warp4JCommandConfiguration.builder().build();
+      var expectedTargets = Set.of(
+          new Target(Platform.LINUX, Architecture.X64),
+          new Target(Platform.LINUX, Architecture.AARCH64),
+          new Target(Platform.MACOS, Architecture.X64),
+          new Target(Platform.MACOS, Architecture.AARCH64),
+          new Target(Platform.WINDOWS, Architecture.X64),
+          new Target(Platform.WINDOWS, Architecture.AARCH64)
+          );
+
+      // When
+      var actualTargets = warp4JCommandConfiguration.getAllSelectedTargets();
+
+      // Then
+      assertThat(actualTargets).isEqualTo(expectedTargets);
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideArchitectureAndPlatforms")
+    void testGetAllSelectedTargets_WhenTargetsAreSelected_ThenTargetsReturned(boolean isLinux, boolean isMacos, boolean isWindows,
+        boolean isLinuxX64, boolean isLinuxAarch64, boolean isMacosX64, boolean isMacosAarch64,
+        boolean isWindows64, boolean isWindowsAarch64, String expectedArchitecture, Set<Target> expectedTargets) {
+      // Given
+      var warp4JCommandConfiguration = Warp4JCommandConfiguration.builder()
+          .linux(isLinux)
+          .linuxAarch64(isLinuxAarch64)
+          .linuxX64(isLinuxX64)
+          .macos(isMacos)
+          .macosAarch64(isMacosAarch64)
+          .macosX64(isMacosX64)
+          .windows(isWindows)
+          .windowsAarch64(isWindowsAarch64)
+          .windowsX64(isWindows64)
+          .architecture(expectedArchitecture)
+          .build();
+
+      // When
+      var actualTargets = warp4JCommandConfiguration.getAllSelectedTargets();
+
+      // Then
+      assertThat(actualTargets).isEqualTo(expectedTargets);
+    }
+
+    static Stream<Arguments> provideArchitectureAndPlatforms() {
+      return Stream.of(
+          Arguments.of(false, false, false, false, false, false, false, false, false, "", Set.of(new Target(Platform.LINUX, Architecture.X64), new Target(Platform.LINUX, Architecture.AARCH64), new Target(Platform.MACOS, Architecture.X64), new Target(Platform.MACOS, Architecture.AARCH64), new Target(Platform.WINDOWS, Architecture.X64),new Target(Platform.WINDOWS, Architecture.AARCH64))),
+          Arguments.of(true, false, false, false, false, false, false, false, false, "", Set.of(new Target(Platform.LINUX, Architecture.X64), new Target(Platform.LINUX, Architecture.AARCH64))),
+          Arguments.of(false, false, false, false, true, false, true, false, false, "", Set.of(new Target(Platform.LINUX, Architecture.AARCH64), new Target(Platform.MACOS, Architecture.AARCH64))),
+          Arguments.of(true, true, true, false, false, false, false, false, false, "x64", Set.of(new Target(Platform.LINUX, Architecture.X64), new Target(Platform.MACOS, Architecture.X64), new Target(Platform.WINDOWS, Architecture.X64))),
+          Arguments.of(true, false, false, false, false, false, false, false, true, "x64", Set.of(new Target(Platform.LINUX, Architecture.X64), new Target(Platform.WINDOWS, Architecture.AARCH64))),
+          Arguments.of(true, true, true, true, false, true, false, true, false, "aarch64", Set.of(new Target(Platform.LINUX, Architecture.X64), new Target(Platform.LINUX, Architecture.AARCH64), new Target(Platform.MACOS, Architecture.X64), new Target(Platform.MACOS, Architecture.AARCH64), new Target(Platform.WINDOWS, Architecture.X64), new Target(Platform.WINDOWS, Architecture.AARCH64)))
+      );
+    }
   }
 }