+++ src/test/java/de/dddns/kirbylink/warp4j/service/FileServiceTest.java
@@ -8,6 +8,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
@@ -24,6 +25,7 @@ import de.dddns.kirbylink.warp4j.config.Warp4JCommand.Warp4JCommandConfiguration
 import de.dddns.kirbylink.warp4j.config.Warp4JConfiguration;
 import de.dddns.kirbylink.warp4j.model.Architecture;
 import de.dddns.kirbylink.warp4j.model.Platform;
+import de.dddns.kirbylink.warp4j.model.Target;
 import de.dddns.kirbylink.warp4j.model.adoptium.v3.VersionData;
 import de.dddns.kirbylink.warp4j.utilities.FileUtilities;
 
@@ -53,8 +55,7 @@ class FileServiceTest {
   @DisplayName("Tests for Windows")
   class TestsForWindows {
 
-    private final Platform platform = Platform.WINDOWS;
-    private final Architecture architecture = Architecture.X64;
+    private final Target target = new Target(Platform.WINDOWS, Architecture.X64);
 
     @Test
     void testExtractJdkAndDeleteCompressedFile_WhenWindows_ThenZipExtractedAndDeleted() {
@@ -77,10 +78,25 @@ class FileServiceTest {
       mockedFileUtilities.when(() -> FileUtilities.optionalExtractedJdkPath(mockedApplicationDataDirectory, "17")).thenReturn(Optional.of(mockedApplicationDataDirectory));
 
       // When
-      var result = fileService.extractJdkAndDeleteCompressedFile(platform, architecture, versionData, mockedApplicationDataDirectory);
+      var result = fileService.extractJdkAndDeleteCompressedFile(target, versionData, mockedApplicationDataDirectory);
 
       // Then
       assertThat(result).isNotNull();
+    }
+
+    @Test
+    void testExtractJdkAndDeleteCompressedFile_WhenWindowsAndException_ThenNullIsReturned() {
+      // Given
+      var mockedApplicationDataDirectory = mock(Path.class);
+      var versionData = mock(VersionData.class);
+      when(mockedApplicationDataDirectory.resolve(anyString())).thenReturn(mockedApplicationDataDirectory);
+      mockedFileUtilities.when(() -> FileUtilities.extractZip(any(), any())).thenThrow(new IOException("No space left. Harddisk full."));
+
+      // When
+      var result = fileService.extractJdkAndDeleteCompressedFile(target, versionData, mockedApplicationDataDirectory);
+
+      // Then
+      assertThat(result).isNull();
     }
 
     @Test
@@ -99,7 +115,7 @@ class FileServiceTest {
       mockedFileUtilities.when(() -> FileUtilities.createZip(mockedBundlePath, mockedCompressedTarget)).then(inv -> null);
 
       // When
-      var result = fileService.compressBundle(platform, architecture, mockedBundlePath);
+      var result = fileService.compressBundle(target, mockedBundlePath);
 
       // Then
       assertThat(result).isTrue();
@@ -122,7 +138,7 @@ class FileServiceTest {
       mockedWarp4JConfiguration.when(() -> Warp4JConfiguration.getLauncherWindows(any(), any(), any(), eq(true))).thenReturn("echo Windows!");
 
       // When
-      var result = fileService.copyJarFileAndCreateLauncherScriptToBundleDirectory(platform, architecture, mockedBundlePath, mockedJarPath, mockedConfig);
+      var result = fileService.copyJarFileAndCreateLauncherScriptToBundleDirectory(target, mockedBundlePath, mockedJarPath, mockedConfig);
 
       // Then
       assertThat(result).isNotNull();
@@ -133,15 +149,14 @@ class FileServiceTest {
   @DisplayName("Tests for Linux")
   class TestsForLinux {
 
-    private final Platform platform = Platform.LINUX;
-    private final Architecture architecture = Architecture.X64;
+    private final Target target = new Target(Platform.LINUX, Architecture.X64);
 
     @Test
     void testExtractJdkAndDeleteCompressedFile_WhenLinux_ThenReturnsExtractedPath() {
       // Given
       var versionData = new VersionData();
       versionData.setMajor(17);
-      versionData.setSemver("17.0.14+7");
+      versionData.setOpenjdkVersion("17.0.14+7");
 
       var mockedAppDataDir = mock(Path.class);
       var mockedJdkRoot = mock(Path.class);
@@ -151,8 +166,8 @@ class FileServiceTest {
 
       when(mockedAppDataDir.resolve(anyString())).thenReturn(mockedAppDataDir);
       when(mockedAppDataDir.resolve("jdk")).thenReturn(mockedJdkRoot);
-      when(mockedJdkRoot.resolve(platform.getValue())).thenReturn(mockedJdkDirectory);
-      when(mockedJdkDirectory.resolve(architecture.getValue())).thenReturn(mockedJdkDirectory);
+      when(mockedJdkRoot.resolve(Platform.LINUX.getValue())).thenReturn(mockedJdkDirectory);
+      when(mockedJdkDirectory.resolve(Architecture.X64.getValue())).thenReturn(mockedJdkDirectory);
       when(mockedJdkDirectory.resolve("jdk.tar.gz")).thenReturn(mockedCompressed);
       when(mockedExtracted.getFileName()).thenReturn(Path.of("jdk-17.0.14+7"));
 
@@ -164,7 +179,7 @@ class FileServiceTest {
       mockedFiles.when(() -> Files.isDirectory(mockedExtracted)).thenReturn(true);
 
       // When
-      var result = fileService.extractJdkAndDeleteCompressedFile(platform, architecture, versionData, mockedAppDataDir);
+      var result = fileService.extractJdkAndDeleteCompressedFile(target, versionData, mockedAppDataDir);
 
       // Then
       assertThat(result).isEqualTo(mockedExtracted);
@@ -186,7 +201,7 @@ class FileServiceTest {
       mockedFileUtilities.when(() -> FileUtilities.createTarGz(mockedBundlePath, mockedCompressedTarget)).then(inv -> null);
 
       // When
-      var result = fileService.compressBundle(platform, architecture, mockedBundlePath);
+      var result = fileService.compressBundle(target, mockedBundlePath);
 
       // Then
       assertThat(result).isTrue();
@@ -211,7 +226,7 @@ class FileServiceTest {
       mockedWarp4JConfiguration.when(() -> Warp4JConfiguration.getLauncherBash(any(), any(), any(), any())).thenReturn("#!/bin/bash\necho Linux");
 
       // When
-      var result = fileService.copyJarFileAndCreateLauncherScriptToBundleDirectory(platform, architecture, mockedBundlePath, mockedJarPath, mockedConfig);
+      var result = fileService.copyJarFileAndCreateLauncherScriptToBundleDirectory(target, mockedBundlePath, mockedJarPath, mockedConfig);
 
       // Then
       assertThat(result).isEqualTo(mockedScriptPath);
@@ -222,8 +237,7 @@ class FileServiceTest {
   @DisplayName("Tests for MacOs")
   class TestsForMacOs {
 
-    private final Platform platform = Platform.MACOS;
-    private final Architecture architecture = Architecture.X64;
+    private final Target target = new Target(Platform.MACOS, Architecture.X64);
 
     @Test
     void testCompressBundle_WhenMac_ThenAppStructureCreated() {
@@ -252,7 +266,7 @@ class FileServiceTest {
       mockedFileUtilities.when(() -> FileUtilities.createTarGz(mockedAppDir, mockedCompressed)).then(inv -> null);
 
       // When
-      var result = fileService.compressBundle(platform, architecture, mockedBundlePath);
+      var result = fileService.compressBundle(target, mockedBundlePath);
 
       // Then
       assertThat(result).isTrue();