+++ src/test/java/de/dddns/kirbylink/warp4j/utilities/FileUtilitiesTest.java
@@ -22,12 +22,16 @@ import java.util.jar.JarOutputStream;
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveOutputStream;
 import org.apache.commons.compress.compressors.gzip.GzipCompressorOutputStream;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.MockedStatic;
+import de.dddns.kirbylink.warp4j.model.Architecture;
 import de.dddns.kirbylink.warp4j.model.JavaVersion;
 import de.dddns.kirbylink.warp4j.model.Platform;
+import de.dddns.kirbylink.warp4j.model.Target;
+import de.dddns.kirbylink.warp4j.model.adoptium.v3.VersionData;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -43,6 +47,12 @@ class FileUtilitiesTest {
   static void setup() throws IOException {
     zipFile = createTestZip(temporaryDirectory);
     tarGzFile = createTestTarGz(temporaryDirectory);
+  }
+
+  @AfterAll
+  static void cleanUp() throws IOException {
+    Files.delete(tarGzFile);
+    Files.delete(zipFile);
   }
 
   @Test
@@ -326,6 +336,55 @@ class FileUtilitiesTest {
     }
   }
 
+  @Test
+  void testCopyJdkToBundleDirectory_WhenJdkExists_ThenJdkPathIsCopied() throws IOException {
+    // Given
+    var target = new Target(Platform.LINUX, Architecture.X64);
+    var applicationDataDirectory = temporaryDirectory.resolve("application-data-folder");
+    var extractedJdkPath = applicationDataDirectory.resolve("jdk").resolve("linux").resolve("x64").resolve("jdk-17.0.15+6/");
+    Files.createDirectories(extractedJdkPath);
+    var versionData = new VersionData();
+    versionData.setMajor(17);
+    var expectedBundleDirectoryPath = applicationDataDirectory.resolve("bundle").resolve("linux").resolve("x64");
+
+    // When
+    var actualBundleDirectoryPath = FileUtilities.copyJdkToBundleDirectory(target, applicationDataDirectory, extractedJdkPath, versionData);
+
+    // Then
+    assertThat(actualBundleDirectoryPath).isEqualTo(expectedBundleDirectoryPath);
+  }
+
+  @Test
+  void testCopyJdkToBundleDirectory_WhenJdkNotExist_ThenJdkPathIsNotCopied() throws IOException {
+    // Given
+    var target = new Target(Platform.LINUX, Architecture.X64);
+    var applicationDataDirectory = temporaryDirectory.resolve("application-data-folder");
+    var versionData = new VersionData();
+    versionData.setMajor(17);
+
+    // When
+    var actualBundleDirectoryPath = FileUtilities.copyJdkToBundleDirectory(target, applicationDataDirectory, null, versionData);
+
+    // Then
+    assertThat(actualBundleDirectoryPath).isNull();
+  }
+
+  @Test
+  void testCalculateSha256Hash_WhenPathExists_ThenHashWillBeCalculated() throws IOException {
+    // Given
+    var temporaryFile = temporaryDirectory.resolve("test.txt");
+    var data = "HelloWorld".getBytes();
+    Files.write(temporaryFile, data);
+    var expectedHash = "872e4e50ce9990d8b041330c47c9ddd11bec6b503ae9386a99da8584e9bb12c4";
+
+    // When
+    var actualHash = FileUtilities.calculateSha256Hash(temporaryFile);
+
+    // Then
+    assertThat(actualHash).isEqualTo(expectedHash);
+
+  }
+
   private void createJarFile(Path jarFileWithModuleInfoClass) throws IOException {
     try (var jarOutputStream = new JarOutputStream(Files.newOutputStream(jarFileWithModuleInfoClass))) {
       jarOutputStream.putNextEntry(new JarEntry("module-info.class"));
@@ -337,7 +396,6 @@ class FileUtilitiesTest {
       jarOutputStream.closeEntry();
     }
   }
-
 
   private static Path createTestZip(Path directory) throws IOException {
     var zipFile = directory.resolve("test.zip");